make loot-picker receive @loot-deck
This means it is no longer "derived," so can be used as-is for further
game state (_e.g._, updating it by removing cards as they are drawn).

The downside is increased complexity: obs-observe! registers arbitrary
callbacks, which we used to update the value of the @loot-deck. But we
have to make sure the observed value is still referenced; _e.g._, the
@deck value from the old code was not a good target for obs-observe!
because changes to it (via @n) somehow disappeared (possibly because it
was garbage-collected?). Additionally, we need to make sure that once we
are done building the @loot-deck, we don't continue to modify it by
modifying the observables generated by make-cards-picker! or by
modifying @random-item?. The good news is the only objects that retain
references to these things do not provide access to them; the bad news
is there is no way to clear the observation call-backs to entirely
prevent it.

As long as it is impossible to re-visit the loot-deck screen once
game-play has started, we should be OK.