use callbacks in player-input-view/s
The design is radically different: in order to get the top-level gui
view working, I need to stick with list-view. But since each sub-view
has call-backs that update a single player observable, I need a list of
observables (and I pair them with their index for the key).

N.B. Defining @player works correctly but may be difficult to reason
about. I like to think of observables as pointers for this part:
@players is effectively a pointer to a list of pointers. Since I'm never
modifying the @players list (only the values pointed to by what's in the
list), I can always peek it when building a view for an entry. IOW, I
should not need to modify what pointers are in the list, so holding onto
one is fine. (If I do make such a modification, I expect the list-view
to deal with it.)

player-input-view is radically simplified by only having observables for
display and using callbacks to allow the parent to decide updates. It
would still be possible to push the callbacks into player-input-view,
but on advice I'm trying to pass fewer observables into a component that
will modify them. I don't mind it for player-input-views because I know
the callbacks will always be the same.

player-info's main simplified: I don't need Debug if I show info and
input at the same time (which is possible due to the observability,
though (cdr (nth (peek @players))) is annoying).