use values in callbacks rather than lists
Since these callbacks have a uniform shape, we don't need the
lists, esp. the overhead of allocating and unpacking them. The lists are
more useful when a single callback needs to handle multiple kinds of
"events" with different shapes/data (though arguably I should be using
structs at that point).