correct multi-monster-picker state
Using `#f` for set meant that the multi-monster-picker (such as from
running `racket main.rkt` or `racket monster-db.rkt`) and not changing
the set would never update the internal state, so the monster group
would not be added.

Because we are in a button handler, we can correctly peek the database,
so we can use an initial set and info. We could alternately set the
initial state in `single-monster-picker` like we do for `info`.