require gui-easy 0.5 for improved conditional view
Thanks to work by @Bogdanp in
https://github.com/Bogdanp/racket-gui-easy/issues/23, I can replace the
`dyn-view` + `obs-combine cons` hack with an actual `cond-view`, now
that `if-view` and descendants are more dynamic and do not force
ahead-of-time evaluation.

Note that `make-monster-group-view` returns to receiving `@ability-decks`
like it effectively did in
https://github.com/benknoble/frosthaven-manager/commit/90114f22112d1353a998252d492fa06b95f92b5f
Then, because `@action` is derived from it (along with `@mg` and
therefore `@e`), the returned view correctly depends on
`@ability-decks`, which was the reason for the `obs-combine cons` hack
in the first place!