doc dev setup
I added the bare minimum reference material as before; this is a helpful
way for me to slowly flesh that out.

I also converted the for-label requires into relative paths. The
originals were generated by some Vimming with the output of `:read !ls`
(or maybe `:read !git ls-files`), and I forgot to convert them to
relative paths.