elements: make right-click popup-menu to any state
Note that `render-popup-menu`
1. requires a renderer, and
2. renders in coordinates specified by the root window of the renderer.
Since the canvases to which we apply the mixins _are_ gui windows
(https://docs.racket-lang.org/gui/windowing-overview.html#%28part._.Core_.Windowing_.Classes%29),
the mouse-click event is reported in coordinates of the canvas. When we
try to render the popup-menu via `(render-popup-menu …)` with a renderer
built from `(render (window … elements-cycler …))`, the menu ends up
being placed at the wrong coordinates.

In a diagram:
------------------------------
|                            |
| p                          |
|   -----   -----   -----    |
|   |   |   |   |   |   |    |
|   |   |   |   |   | x |    |
|   -----   -----   -----    |
|     A       B       C      |
|                            |
------------------------------

The 'x' marks the click. We want the popup to appear there, but the
coordinates we are given are in terms of the canvas 'C' (roughly (1,1)
here). When rendered on the outer window, that shows the popup-menu at
position 'p'!

We could borrow the logic of `render-popup-menu` from
`racket/gui/easy/private/renderer`, but swapping in the canvas window
(_i.e._, `this`). That way, the coordinates are specified appropriately.
This is done by something of a hack: the `renderer%` class is not
available publicly from `racket/gui/easy` or from its internal private
modules, so to access it we use Racket's namespace reflection
capabilities. Code:
```
(define (handle-element-clicks cycle-element @state)
  (define renderer%
    (let ([mod 'racket/gui/easy/private/renderer])
      (dynamic-require mod #f)
      (define ns (module->namespace mod))
      (eval 'renderer% ns)))
  (define (gui:popup-menu pum)
    (define r (new renderer% [tree pum]))
    (begin0 (send r render #f)
      (send r destroy)))
  (mixin (canvas<%>) ()
    (super-new)
    (define/override (on-event e)
      (case (send e get-event-type)
        [(left-down) (cycle-element)]
        [(right-down)
         (define pum
           (gui:popup-menu
             (popup-menu
               (menu
                 "Transition to…"
                 (menu-item "Infused" (λ () (:= @state 'infused)))
                 (menu-item "Waning" (λ () (:= @state 'waning)))
                 (menu-item "Unfused" (λ () (:= @state 'unfused)))))))
         (send this popup-menu pum (send e get-x) (send e get-y))]))))
```

If there were a way to translate the coordinates of the canvas to
coordinates of the parent window, we could use
`render-popup-menu` with the renderer for the parent window.

Fortunately, there is such a way! See `translate-to-top-coords`.

Notably, the helper omits the "top" frame because
1. The pum is rendered in the top's coords, and
2. The actual top-level widget's `get-x`/`get-y` methods return
   screen-level coordinates (? "… in it's parent's coordinate system",
   but the parent is `#f`…).