introduce current-renderer parameter
There are bugs lurking here: the state is odd, particularly how it can
interact with GUI callbacks. It's possible to forget to set it in a
`main` that needs it.

I've considered wrapping every render with `(thread (thunk …))` so I can
unconditionally set `current-renderer` in that thread, but that hasn't
worked very well in my testing so far.