compute version at compile (expansion) time
Collections are gone when bundling with `raco exe` and `raco
distribute`, no matter what hacky flags I try to include them with.
There is an `#%info-lookup` provided by frosthaven-manager/info, but it
is undocumented and not intended to be public at the moment.

Since the modules are expanded by the bundling process, this embeds the
version directly in the code at build time. Collection lookups are not
later required.