try setup after install
The v0.5 artifacts exhibit a weird issue that I also had locally, but
only while using outdated code: the following commands made the issue
disappear:
```
find . -type d -name compiled -delete -print
make setup
make FrosthavenManager.app/Contents/MacOS/FrosthavenManager
```

The issue is that on the `choose-monster-db-view` screen, both bestiary
buttons failed and, if you ran the program in a terminal, logged the
following:
```
collection-path: collection not found
  collection: "frosthaven-manager"
  in collection directories:
  context...:
   .../private/collect.rkt:11:53: fail
   .../setup/getinfo.rkt:17:0: get-info
   .../private/arrow-val-first.rkt:555:3
   body of '#%embedded:frosthaven-manager/gui/common-menu:
   .../racket/lazy-require.rkt:89:13: get-sym
   .../racket/promise.rkt:68:10
   .../racket/promise.rkt:48:2
   .../racket/lazy-require.rkt:109:6
   .../data/monad.rkt:66:23
   .../data/monad.rkt:31:0
   .../megaparsack/base.rkt:134:6
   [repeats 1 more time]
   .../match/compiler.rkt:548:40: f236
   .../megaparsack/base.rkt:134:6
   [repeats 1 more time]
   .../megaparsack/base.rkt:302:3
   ...
```

This proved reliably repeatable locally (until after the `find`/`make`)
and in the macOS artifact for v0.5. But it also proved indecipherable.
Neither of the two (!) code paths in the trace or in the buttons has
anything to do with any of the others.