always construct sub-views of dynamic views
In b4208472f5da847dd19435d743d62743c45b4264, I fell victim to a trap of
`dyn-view`/`if-view` and related views: the child views are destroyed
and recreated as needed when switching between the dynamic components. I
thought this might be an issue when I wrote the code but didn't notice
anything… until investigating
https://github.com/benknoble/frosthaven-manager/issues/43.

I tried _everything_: logging via obs-observe! makes it 100% clear that
the ability card is correctly being pulled, but the view isn't updating.
Process of elimination reveals this to be the culprit: changing the
child-views to be functions (so that fresh views are constructed each
time) completely solves the issue.

Curse you mutable state!

Close https://github.com/benknoble/frosthaven-manager/issues/43