don't show move stats anymore
We'll keep the attack stats, because otherwise the default height on the
ability card area is a bit small (and it's nice as a player to gauge
attacks). But with the inline calculations, neither is _really_ needed.
Then again, neither is Max HP, technically. But we're keeping that, too,
obviously.