Fix the "weird bug"
This reverts commit c80bf39ebce3662429f2820feb46b2dc3e84f1ef.

The problem: not duplicate titles, but duplicate ability text (since
each produced identical objects for the `list-view`).

The solution: pair each ability text with a gensym, which is guaranteed
to be unique and should not have a performance or memory impact (since
they are not interned and therefore are eligible for GC). My testing did
not show any issues with this approach.

Notably, using `(map (flow (cons (gensym))))` didn't work; they would
all get the same gensym. This is because `(flow (cons (gensym)))`
evaluates to `(curryr cons (gensym))`, and `curryr is _eager_. In other
words, flows are not supposed to have side-effects outside of `ε`, which
we know.