remove gensym-key hack for monster abilities
The switch to observable-view simplified some code which was previously
very observable driven. It also revealed an odd error in the "else" case
for aoe pict paths: when the @ability-card changed so that @base was
effectively wrong, the console output messages (presumably because the
@pict observable hadn't been garbage collected yet). This happens when
an AoE ability is "cleared" by clicking "Next Round."

As a solution, build in more checking to the @pict value.