web: handle mixed request bindings better
The existing match clauses in, for example, -do-monster would allow
{mgid mn} to bind to {<actual-id> #f}, which we don't want. The _intent_
is to do nothing via (void) when _any_ of the values is missing.

For current code that check [#f (void)] before [<some-val> <action>],
there is no risk. But where there are multiple values, the current
expressions check "both #f" and allow all (three) other combinatorial
cases. With more values, it would get only worse.

New code explicitly allows non-#f values, and throws the rest away. This
is smarter, and we'll change the single-value code to do the same for
consistency.