/actions/player/loot: return JSON information
This could be used for an alert, for example.

One caveat is that multiple simultaneous requests may get the wrong
return values since access to (s) isn't locked: both might return the
same card because each evaluates the `(define card …)` before either
evaluates the `(do …)`. Fortunately, it doesn't affect correctness of
the underlying state (or Rewards page or Rewards GUI), only the possible
messages shown later. This race is unlikely to happen in "live" play,
since each character takes their turn individually and kind players are
unlikely to loot out of turn.

There's a race with num-players, too, but that's unlikely to change at
this point.