web: smash some things inline on larger screens
Note the weird bug on iPhone: rotating to landscape triggers the media
query, but scales some text weirdly. The linked SO Q&A describes how to
disable that without also disabling zoom. If it proves to be an issue on
other browsers, the same link mentions what appears to be a Microsoft
property equivalent `-ms-text-size-adjust`.