repair defns deserialization
The refactoring in ceb011b (defns: extract new modules, 2023-11-18) did
not reprovide deserialization information for structs declared with
serializable-struct (because that information is only automatically
provided in a submodule).

Now that I've repaired this, it seems odd that manager.rkt should work
to deserialize state values (since it doesn't include the submodules for
serializable-structs in manager/*.rkt).