extract sample data
It's really hard to figure out why (module+ test (require testdata)) or
(module* test racket (require (submod "..") testdata)) create module
cycles; the short version according to mflatt is that submodules
(possibly of any kind?) do have a dependency edge on enclosing modules
in the module graph. The edge is not that of a require, but that of
macro expansion: to produce the expansion for the whole module, we must
macro expand submodules (at which point require cycles are detected).

Matthew says that removing the dependency edge is possible in principle,
but requires generalizing the protocols from the expander to zo files to
raco make.

Matthew also notes the only special case of making a module partially
available to others is during the compilation of submodules, where the
enclosing module is partially available. The expander can pause
expanding one module to expand another but not otherwise make the paused
module available.

Anyway, the dynamic-require seems to fix it for now.