transition: handle n-players < n-creatures
This fixes the BUG from the previous commit: namely, when a player is
removed, we discard all player information and start over. This is
better than guessing which player(s) to keep.

Please do NOT interleave the := and the cond to
```
(:= (state-@creatures) (cond …))
```
The cond explicitly avoids having any side effects when n-players =
n-creatures.