defns/monsters: fix tests for rich text model
diff best viewed with: --ignore-all-space

Now that monster-ability-ability->rich-text actual returns more values
in the model, the tests fail. The values changed (see the update to the
"Attack, X" test); moreover, we need the model's model-equal? function
to truly compare results.

This commit fixes the tests to use a new test-model-equal? form which
delegates (through a binary checker) to model-equal?. Now the tests pass
again.