icons: substitute all occurrences
The new test case demonstrates the need for match-loop (which could
possibly use only-on-text, too).

In order to make the test pass, the expected and actual results need to
be consistent in whether or not they use scale-icon; a previous test
with the push icon unscaled confused me. This commit also uniformly
scales the icons in the rich text conversion and in tests to avoid that.