curlique: add experimental Qi syntax
This library provides a Qi syntax that makes {} automatically compile to
flows. The ~>, ~>>, and switch forms get special syntax to avoid the
wrapping () within {}; in other words, you can write
    {~> (-< add1 sub1) /}
instead of
    {(~> (-< add1 sub1) /)}
(though the latter should also work).

This commit does not document curlique (yet): I intend for it to bake
here before becoming its own package.

This commit also does not add curlique to any of the actual code (yet).
That will be forthcoming.