extract define-error-text
Getting this macro right (with good error reporting and small code size
for a macro-generating macro) was fun, because I almost had it on the
first try! I mistakenly used internal-definition-context? instead of
syntax-local-context, but now I know about that. And I don't need to
smuggle the macro name into the error message; it's present already.
Not syntax-quoting the @error-text identifier did bite me but in
retrospect the #' is definitely needed.

Close https://github.com/benknoble/frosthaven-manager/issues/82