reorder effects in reshuffle-modifier-deck
Using the "Shuffle Deck" button in modify-monster-deck-menu-item
triggers reshuffle-modifier-deck in the manager; when the deck is
shuffled _before_ the discard pile is emptied, the observable

  (obs-combine append
               (state-@monster-modifier-deck s)
               (state-@monster-discard s))

temporarily has more cards than it should (it contains the discard pile
twice, effectively). This causes absent-from-modifier-deck to throw an
exception, though it is harmless.

Emptying the discard _first_ means the combined observable temporarily
has less cards, which is fine as far as absent-from-modifier-deck is
concerned.

Close https://github.com/benknoble/frosthaven-manager/issues/67