monster-group-view: support arbitrary update functions
This provides a flexible protocol, though it is hard for a consumer to
predict when it will be triggered. Fortunately, the typical event
handler is something like

    #:on-update (λ (f) (<@ @monster-group f))

That is, apply the emitted event to update the monster group.

As a demonstration, it simplifies the mass condition handler: the state
change is now a single step (and thus a single undo) and is more
performant than a series of `on-condition` events.