experimental: push undo in a thread
This makes changing the state faster, as observer functions yield
(return) faster: thread mailboxes are buffered so that thread-send is
fast and non-blocking.

I haven't noticed any bugs with Undo, but some behavior (like multiple
undos after next-round) is still unintuitive. I believe this is because
of many changes in some cases that don't cause meaningful changes to the
state but still call push-state, so let's investigate de-duplicating
runs of equal states.