undo: each play-view gets its own
This simplifies how far back the undo history goes (to the moment play
started) and makes sure to include the first "play" state.

In the old version, the transition to-play didn't take any actions that
recorded an undo state. When the first action taken on the play-view
does record an undo state, the initial play-view state is lost.

The effects of this will be even more interesting once
https://github.com/benknoble/frosthaven-manager/issues/88 is
implemented.