fix: make loot deck count accurate
When adding a loot deck with special cards, those cards are not
accounted for in update-loot-deck-and-num-loot-cards, so the loot button
might show (5/3), which is confusing.

Instead, let update-loot-deck-and-num-loot-cards build an accurate count
of non-special cards while manipulating the loot deck, but finalize the
count when constructing the deck by using the actual deck.