fix: render loot deck preview in a dialog
Something about this particular widget causes the following exception to
be raised when loot cards are _assigned_ (drawing them is not
sufficient) after using the previewer:

    queue-callback: eventspace is shutdown: #<eventspace>
      context...:
       /Applications/Racket v8.14/share/pkgs/gui-lib/mred/private/wxwindow.rkt:130:10: enable method in wx-make-window%
       /Applications/Racket v8.14/collects/ffi/unsafe/atomic.rkt:73:13
       /Users/Knoble/Library/Racket/8.14/pkgs/gui-easy-lib/gui/easy/private/view/container.rkt:22:4: update-children method in container%
       [repeats 2 more times]
       /Applications/Racket v8.14/share/pkgs/gui-lib/mred/private/wx/common/queue.rkt:436:6
       /Applications/Racket v8.14/share/pkgs/gui-lib/mred/private/wx/common/queue.rkt:487:32
       /Applications/Racket v8.14/share/pkgs/gui-lib/mred/private/wx/common/queue.rkt:639:3

I think it has something to do with observable events still being sent
to the GUI even though the eventspace is shutdown, but I haven't figured
out why that is. Making it a dialog instead (so not using an eventspace)
got rid of the exception, but may be hiding a problem with retaining a
non-destroyed widget in the GUI tree. Unclear: running with errortrace
doesn't reveal any more information, unfortunately (I'd probably have to
attach errortrace to the frameworks in place, too?).