gui/main: make sure there's a bestiary
When adding a new monster starting from 'play mode, there may not yet be
a bestiary. This breaks Cmd+N and the menu item.

We could fix it by setting the default bestiary to default-monster-db in
manager/state's constructor for the state. Instead, have the "New
Monster" dialog request a bestiary when needed, deferring the choice and
also teaching the user there is a choice.