fix: make transitions idempotent
This helps avoid (but not entirely prevent) a situation in which the
next-round (or draw-abilities) transition is called more than once in a
row: in the wrong state, these actions now do nothing.

With the web-server running, it's possible for 2 players to click "Draw
Abilities" close enough together than both progress-game calls attempt
to trigger the same transition function. Since we don't (yet) serialize
server callbacks, this creates a further problem: one execution of the
transition may occur at the same time as another. Not only is that
difficult to program against, but it typically creates such a confusing
game state that it is difficult to undo and continue playing.

This does not cure the above confusion, but it may help make it less
likely.

Best-viewed-with: --ignore-space-change