server: serialize all callbacks
Continuing the issue from 64e77d7 (fix: make transitions idempotent,
2024-10-04), this makes _all_ callbacks from the server execute in the
same event queue as the GUI (though server callbacks have higher
priority than anything else bar callbacks from gui/render).

This has the effect of serializing all server actions from all players,
which is sort of like sequential consistency. In other words, it may be
_too_ strong for this particular problem in that it hurts responsiveness
when multiple players are tapping actions, but that remains to be seen
in a live game.

A slightly weaker possibility is to include the queue-callback change in
progress-game's do-action, so that the callback is installed eventually
(from the response-handler's thread, where all the events used to run,
but with the GUI's eventspace as normal): when it runs, it would queue
the callback into the appropriate queue, but consider processing other
callbacks from other response-handler threads concurrently.