make progress-game execution consistent
Now that we serialize server event execution (013778f (server: serialize
all callbacks, 2024-10-04)), the double-player-tap on "Draw Abilities"
would invoke the 2 progress-game callbacks in sequence, one completing
entirely before the other. The callback, though, has to decide which
transition to trigger, and it did so based on the current state.

Instead, determine which callback to trigger at the time the thunk is
created. Thus, instead of the double-player-tap accidentally skipping a
whole round, both taps trigger draw-abilites, the second evaluation of
which does nothing.