fix: restore server action functionality
After 013778f (server: serialize all callbacks, 2024-10-04), server
action callbacks execute after the request handler is dead (and the
parameterization is reset), so they all see #f for (s). This effectively
breaks the server.

Now: the handler accepts events that are invoked with a state parameter
so that it may pass the appropriate state. The do macro changes to
accommodate this, unhygienically binding s inside the body.

Now that things are working again, web-draw modifier is broken: it
creates an alert response from the current modifier, but the modifier
draw hasn't happened yet (it's only been queued). Since trying to
predict the modifier is hard, I will either alert all players of the
results (see the `(alert …) SSE) [easy] or try using continuations to
(re-)respond with a modifier after completing the draw [hard].