fix: work harder to make sure state is available in server callbacks
See also b612c58 (fix: restore server action functionality, 2024-10-04):
this handler used (s) to retrieve the state but as noted it is often #f:
this caused a recent but hard to reproduce bug where state-@env raises a
contract violation.

To fix it, thread state from the do-monster-group/mgid helper into the
action, which is now tagged with needs-state. The tag allows
distinguishing regular actions from actions that require state and
produce regular actions; this way, do-monster-group/n and its many
callers are unaffected.

I think increment-monster-hp is the only caller of
do-monster-group/mgid, so this fix might be extra code for a single
caller. Only time will tell if it ends up being needed elsewhere.