server: provide feedback on curse/bless monster
Otherwise, the user has no way of knowing that the button press was
registered and handled, short of the player running the desktop app
confirming. This is slow and error-prone.

As a bonus, this quickly answers the question of "how many curses are in
the monster deck," which my players and I frequently ask (because we
don't want to add 5 curses with just 1 left).