syntax: synthesize fewer identifiers
These identifiers do _not_ need to be synthesized in order to be visible
in provide specifications, contrary to my earlier beliefs.

However, as a result, the input #'info-db to make-dbs is no longer
original in the sense of coming from the original module's syntax: this
is a problem because it means that the define-runtime-path form gets the
wrong context, causing the AoE modules to not be found at runtime.

Perhaps there is a way to use imports, infos, or actions as appropriate
sources, but using #'(imports ...) as the context and source for
datum->syntax also failed. Instead, pass an original syntax that can be
used for this information (and document it). Note that this requires
quasisyntax in order to embed this-syntax from syntax-parse.