fix: call monster-group-monsters with a valid value
In a recent game, I got this stacktrace:

    monster-group-monsters: contract violation
      expected: monster-group?
      given: #f
      in: the 1st argument of
          (-> monster-group? (listof monster?))
      contract from:
          <pkgs>/frosthaven-manager/defns/monsters.rkt
      blaming: <pkgs>/frosthaven-manager/manager/state.rkt
       (assuming the contract is correct)
      at: <pkgs>/frosthaven-manager/defns/monsters.rkt:91:21
      context...:
       /Applications/Racket v8.15/collects/racket/contract/private/blame.rkt:350:0: raise-blame-error
       /Applications/Racket v8.15/collects/racket/contract/private/arrow-higher-order.rkt:383:33
       /Applications/Racket v8.15/collects/racket/private/list.rkt:391:30
       /Users/Knoble/code/frosthaven-manager/manager/state.rkt:434:0: kill-monster
       /Applications/Racket v8.15/share/pkgs/gui-lib/mred/private/wx/common/queue.rkt:436:6
       /Applications/Racket v8.15/share/pkgs/gui-lib/mred/private/wx/common/queue.rkt:487:32
       /Applications/Racket v8.15/share/pkgs/gui-lib/mred/private/wx/common/queue.rkt:639:3
    2024-11-07 21:33:51.236 racket[70979:1251615] Warning: Window move completed without beginning

I'm pretty sure the window move warning is unrelated; I'm also not sure
how the `when` block was executed without the `<@` having set `g` to
contain a monster group (unless… there were none? but then how was
kill-monster called? maybe a button was double-clicked and too slow to
respond?). While the issue didn't cause real game problems, fix it to
prevent future problems: check that `(unbox g)` is not #f! By
monster-group-remove's contract, it must be a monster group.