bestiary: expand to compile-time substitution of AoEs
Use syntax-local-lift-require to embed require forms for each found AoE
spec, obviating 600a2eb (bestiary: require AoE forms in expansion,
2024-11-11). The current implementation generates one require per aoe()
form, which is not particularly efficient if the same AoE is used
multiple times in the same module. That can be a common occurrence. I
have not, however, benchmarked the overhead relative to trying to
intelligently collapse requires by processing the whole module like
find-imports does, and I expected we can get some help from Racket's
module semantics (requiring the same module multiple times into a single
client should be collapsible to a single require, right?).

As a result of the new expansion, fix and test some bugs from 78f3078
(prep: extend monster-ability contract to accept picts, 2024-11-16). In
particular:

- When the expander produces (list x y z) for an ability-parts, only x
  should get a bullet (assuming its a string). If there's no leading
  string, insert the bullet.
- We need to filter out empty strings (which may result from the
  expander's compile-time computations).
- We do not (currently) smuggle newlines around AoEs through the
  expander, so we should add them here. Given this change, it may not be
  worth adding newlines to the expander code.

------------------------------------------------------------------------

As for implementation, there are 2 important details.

1. We interpose with a new macro in the make-dbs expansion result at the
   monster-ability struct. We want to compute over the
   #s(monster-ability …) prefab at compile-time.
2. The new macro, process-aoes, computes the resulting syntax object by
   examining the prefab datum at compile time. After extracting the
   pieces, it calls splice-aoes at compile time on each part. The result
   is shoved into the monster-ability constructor, which otherwise
   preserves the input data. The splice-aoes algorithm is essentially
   the following:
    a. Collect a map of aoe specs (the bit inside "aoe()") to generated
       temporaries. It is important to keep the temporaries as syntax
       objects.
    b. Then, perform a substitution using this map on all the parts,
       which proceeds in a fixed-point loop similar to that of
       part->rich-text [2]. The resulting replacement needs to be a
       datum that will eventually be converted to syntax, so the
       inserted temporary expression is a list like what '(make-aoe)
       evaluates to. Writing `(,(hash-ref …)) would produce the same
       result as the construction in this commit.
    c. Then, compute the expression that will be the replacement for
       each individual monster-ability-action piece: namely, we compute
       (list part ...) since we are sticking the result in a non-quoted
       context. Otherwise we'd insert (part ...) which won't work.
    d. Bend hygiene: make the literal data appear as though it came from
       the input context (action). Syntax objects (like the temporaries
       and #'list) are unmodified. There's no bindings in the literal
       data that require actual hygiene modifications, but explicitly
       converting seems like good practice?
    e. Finally, lift requires for each aoe-spec in the collected map,
       where the raw #%require spec renames the import aoe into the
       generated temporary.

Making hygiene work correctly here is tricky because we play with raw
data and convert back to syntax [1].

[1]: In an uncommitted attempt, I wanted to expand to a quasiquoted
prefab struct with unquoted computations for some fields. I couldn't
make this work because the scopes on the qq and the unquote were
different, turning `#s(S x ,y z) into '#s(S x ,y z), which doesn't work.
Rather than fix the scopes, I expand to a constructor call: this
necessitates the extra (list) constructors.

[2]: I've never explained the loop process from 7664c99 (icons:
substitute all occurrences, 2024-04-23) in detail. The essential idea is
to compose functions (-> part (listof part)) with append-map so that
turn (listof part) into (listof part). But we iterate each function
until it no longer applies (using match-loop in the original commit or a
single hand-written implementation here) to perform the substitutions on
_all_ instances in the part. We also apply each function only-on-text;
this version takes extra arguments. Both are specific to the append-map
loop---they know to return (list input) when not applying the
transformation, which then gets spliced into the append-map result. A
similar wrapper is not applied to the result of the transformation
because it already returns a list. After ripping out no-longer-needed
code, this demands a consolidation… I've even considered running more of
the rich-text conversion at compile time?

Ref: https://github.com/benknoble/frosthaven-manager/issues/108