bestiary: rewind prototype AoE require forms in expansions
This partially reverts commit 600a2eb (bestiary: require AoE forms in
expansion, 2024-11-11); specifically, changes to testfiles/ are kept
along with the modifications to syntax/monsters.rkt's requires, but the
use of find-imports to embed more require specs is reverted.

This is a "good" revert: the original commit was not buggy but rather
incomplete relative to the desired end-state.

Ref: https://github.com/benknoble/frosthaven-manager/issues/108