bestiary: rip out old location and runtime splice-aoe code
Now that we embed AoEs in the expansion via requires since a98825e
(bestiary: expand to compile-time substitution of AoEs, 2024-11-18), we
no longer need to embed define-runtime-path and locations in the
actions. Woohoo!

This is not a breaking change but does require recompiling data files
(save files should continue working).

This rips code out from commits like
- 20b5075 (simplify adding location in syntax/monsters, 2023-03-20)
- 4b5848b (include original location in ability card, 2023-03-20)
- 4656028 (syntax: synthesize fewer identifiers, 2024-10-27)
- parts of 1e8d178 (add GUI support for AoE patterns, 2023-03-20) which
  got converted to d283be6 (combine images and text for monster
  abilities, 2024-02-05).

Geez, 20 months in the making…

Ref: https://github.com/benknoble/frosthaven-manager/issues/108