update-all-players: preserve cons cells
See previous commits. This function is only used to clear initiative
right now, but since players are mostly (always?) kept at the front of
the creatures list, this could help a bit with sharing and
transitioning since we'll stop the loop as early as we can.

It's probably *not* worth making the same change to
update-all-monster-groups because the monster groups tend to be the tail
(defeating sharing) and the extra bookkeeping to track number of monster
groups (like we do for players to allow us to bail early) probably isn't
going to let us bail earlier than the end of the list.