pp: repair ability printing
After commit 78f3078 (prep: extend monster-ability contract to accept
picts, 2024-11-16) which changed the monster ability representation,
pretty-ability's assumption about the input data was no longer valid:
instead of a list of strings, it now received a nested list of strings.

As a symptom of the issue, formatting ability decks resulted in extra
parentheses wrapping the game text, like

    ["double-shot" 25 {("attack +2, range 5") ("attack +2, range 3, +1 if same target")}]

which obviously doesn't parse back using parsers/monster.rkt. This is
worse than an error! Only attaching lots of internal contracts could
have caught this, however.

As a fix, we _could_ invoke flatten on the list of abilities while
keeping the rest of the code intact; that is, the following patch
_should_ suffice:

-        (pretty-list (map pretty-string abilities)))))
+        (pretty-list (map pretty-string (flatten abilities))))))

However, the contract on monster-ability values does indicate it could
contain a pict value. This won't happen when invoked in a "normal" way
(by composition with parse-bestiary, which produces only strings, such
as in the main submodule), but might occur if someone wants to program
against the pretty-printer. Explicitly error in such a case.